Ur gunna luv this one.
This now is more impossible to read.
Basically everything is attempted to be forced into a F64,this speeds up math.
AddMethod uses a special FramePtr name which is filled in later.
Static typing is there but not really used. Ill elloborate on this more over the weekend with "Static-Typed-Frogs-Code".

Nrootconauto