Preliminary Quests.
Run `QuestEditor.HC` to edit a quest. I still have lots to do.
Press 'q' in the level editor to add a quest to a thing.
Use the Mouth item in the game to talk to a quest person.

TODO
Make a ->next thing for CEliza so you can have quests inherit from each other.
This would allow to save space and time by having common repsonses like "hello" be stored in a "generic" quest.

Nrootconauto